+++ huggingface/transformers/image-text-to-text/gemma3/gemma-3-4b/compile.py
@@ -12,7 +12,7 @@ def main():
         export=True,
         rbln_config={
             "language_model": {
-                "tensor_parallel_size": 8,
+                "num_devices": 8,
                 "kvcache_partition_len": 16_384,
                 "use_inputs_embeds": True,
             }