+++ huggingface/transformers/image-text-to-text/paligemma2/compile.py
@@ -1,18 +1,18 @@
 import os
 
-from optimum.rbln import RBLNAutoModelForVision2Seq
+from optimum.rbln import RBLNAutoModelForImageTextToText
 
 
 def main():
     model_id = "google/paligemma2-3b-mix-224"
-    model = RBLNAutoModelForVision2Seq.from_pretrained(
+    model = RBLNAutoModelForImageTextToText.from_pretrained(
         model_id,
         export=True,
         rbln_config={
             "language_model": {
                 "batch_size": 1,
                 "max_seq_len": 8192,  # default "max_position_embeddings"
-                "tensor_parallel_size": 4,
+                "num_devices": 4,
                 "prefill_chunk_size": 8192,
             },
         },