+++ huggingface/transformers/sentence-similarity/qwen3-reranker-0.6b/compile.py
@@ -12,7 +12,7 @@ def main():
         export=True,  # export a PyTorch model to RBLN model with optimum
         rbln_batch_size=2,
         rbln_max_seq_len=40_960,  # default "max_position_embeddings"
-        rbln_tensor_parallel_size=1,
+        rbln_num_devices=1,
         rbln_attn_impl="flash_attn",
         rbln_kvcache_partition_len=8192,  # Length of KV cache partitions for flash attention
         rbln_phases=["prefill"],