+++ huggingface/transformers/text2text-generation/phi/phi-2/inference.py
@@ -47,10 +47,8 @@ def main():
     )
 
     # Show text and result
-    print("--- text ---")
-    print(args.text)
-    print("--- Result ---")
-    print(generated_texts)
+    print(f"Text: {args.text}")
+    print(f"Result: {generated_texts}")
 
 
 if __name__ == "__main__":