+++ huggingface/transformers/text2text-generation/qwen/deepseek-r1-distill-qwen-32b/compile.py
@@ -12,7 +12,7 @@ def main():
         export=True,  # export a PyTorch model to RBLN model with optimum
         rbln_batch_size=1,
         rbln_max_seq_len=131_072,  # default "max_position_embeddings"
-        rbln_tensor_parallel_size=16,
+        rbln_num_devices=16,
         rbln_attn_impl="flash_attn",
         rbln_kvcache_partition_len=16_384,  # Length of KV cache partitions for flash attention
     )