+++ serving/triton_inference_server/llama3.1-8b/setup.sh
@@ -13,7 +13,7 @@ echo "=================================================================="
 DIR="vllm_backend"
 GIT_REPO="https://github.com/triton-inference-server/${DIR}.git"
 if [ ! -d ${DIR} ]; then
-  git clone ${GIT_REPO} -b r25.08
+  git clone ${GIT_REPO} -b r26.06
   if [ $? -ne 0 ]; then
     echo "Error while cloning git repository."
     popd